Move range partitioning SLT fixture to module
